Salary готово но нужно чтобы участи считало функция

+++ sms/Rawmaterials/models.py
@@ -3,8 +3,9 @@ from unitofmaesurement.models import Unit
 class RawMaterials(models.Model):
     Name = models.CharField(max_length=255)
     Unit_of_measurement_id = models.ForeignKey(Unit, on_delete=models.CASCADE, db_column='Unit_of_measurement_id')
+
     Quantity = models.FloatField()
-    Amount = models.DecimalField(max_digits=19, decimal_places=2)
+    Amount = models.FloatField()
 
     class Meta:
         db_table = 'RawMaterials'
